payment/requirements.txt
fastapi[standard]==0.124.0
uvicorn[standard]==0.38.0
uvloop==0.21.0
sqlmodel==0.0.27
asyncpg==0.31.0
alembic==1.17.2
pydantic-settings==2.12.0
psycopg==3.3.2
pyjwt==2.10.1
aiokafka==0.12.0
httpx==0.28.1
structlog==25.3.0
cryptography==44.0.2
